+++ lib/json.ex
@@ -3,6 +3,8 @@ defmodule JSON do
   Provides a RFC 7159, ECMA 404, and JSONTestSuite compliant JSON Encoder / Decoder
   """
 
+  require Logger
+
   alias JSON.Encoder, as: Encoder
   alias JSON.Decoder, as: Decoder
 
@@ -51,7 +53,22 @@ defmodule JSON do
   @spec decode(bitstring) :: {atom, term}
   @spec decode(charlist) :: {atom, term}
   def decode(bitstring_or_char_list) do
-    Decoder.decode(bitstring_or_char_list)
+    bitstring_or_char_list |>
+      Decoder.decode() |>
+      case  do
+       res = {:ok, value} ->
+         Logger.debug("#{__MODULE__}.decode(#{inspect bitstring_or_char_list}} was sucesfull: #{inspect res}")
+         res
+       e = {:error, {:unexpected_token, tok}} ->
+         Logger.error("#{__MODULE__}.decode!(#{inspect bitstring_or_char_list}} unexpected token #{tok}")
+         e
+       e = {:error, :unexpected_end_of_buffer} ->
+         Logger.error("#{__MODULE__}.decode!(#{inspect bitstring_or_char_list}} end of buffer")
+         e
+       e ->
+         Logger.error("#{__MODULE__}.decode!(#{inspect bitstring_or_char_list}} an unknown problem occurred #{inspect e}")
+         e
+     end
   end
 
   @doc """
@@ -67,15 +84,16 @@ defmodule JSON do
   def decode!(bitstring_or_char_list) do
     case Decoder.decode(bitstring_or_char_list) do
       {:ok, value} ->
+        Logger.debug("#{__MODULE__}.decode!(#{inspect bitstring_or_char_list}} was sucesfull: #{inspect value}")
         value
-
       {:error, {:unexpected_token, tok}} ->
+        Logger.error("#{__MODULE__}.decode!(#{inspect bitstring_or_char_list}} unexpected token #{tok}")
         raise JSON.Decoder.UnexpectedTokenError, token: tok
-
       {:error, :unexpected_end_of_buffer} ->
+        Logger.error("#{__MODULE__}.decode!(#{inspect bitstring_or_char_list}} end of buffer")
         raise JSON.Decoder.UnexpectedEndOfBufferError
-
-      _ ->
+      e ->
+        Logger.error("#{__MODULE__}.decode!(#{inspect bitstring_or_char_list}} an unknown problem occurred #{inspect e}")
         raise JSON.Decoder.Error
     end
   end
