+++ lib/json/parser/array.ex
@@ -3,9 +3,6 @@ defmodule JSON.Parser.Array do
   Implements a JSON Array Parser for Bitstring values
   """
 
-  alias JSON.Parser, as: Parser
-  import Parser, only: [trim: 1]
-
   @doc """
   parses a valid JSON array value, returns its elixir list representation
 
@@ -30,7 +27,7 @@ defmodule JSON.Parser.Array do
       {:ok, ["foo", 1, 2, 1.5], " lala"}
   """
   def parse(<<?[, rest::binary>>) do
-    rest |> trim |> parse_array_contents
+    rest |> String.trim() |> parse_array_contents()
   end
 
   def parse(<<>>), do: {:error, :unexpected_end_of_buffer}
@@ -46,20 +43,18 @@ defmodule JSON.Parser.Array do
   defp parse_array_contents(_, <<>>), do: {:error, :unexpected_end_of_buffer}
 
   defp parse_array_contents(acc, json) do
-    case json |> trim |> Parser.parse() do
+    case json |> String.trim()|> JSON.Parser.parse() do
       {:error, error_info} ->
         {:error, error_info}
-
       {:ok, value, after_value} ->
-        after_value = trim(after_value)
-
-        case after_value do
-          <<?,, after_comma::binary>> ->
-            parse_array_contents([value | acc], trim(after_comma))
-
-          _ ->
-            parse_array_contents([value | acc], after_value)
-        end
+        after_value |>
+          String.trim() |>
+          case  do
+            <<?,, after_comma::binary>> ->
+              parse_array_contents([value | acc], String.trim(after_comma))
+            _ ->
+              parse_array_contents([value | acc], after_value)
+          end
     end
   end
 end
