+++ lib/json/parser/record/chunk.ex
@@ -1,0 +1,4 @@
+defmodule JSON.Parser.Record.Chunk do
+  @moduledoc false
+  defstruct chunk: "", chunk_id: 0
+end
